refactor: hook 및 util 분리(#53)
