+++ src/hooks/timetable/useImageUpload.ts
@@ -1,0 +1,46 @@
+import { useState, useEffect } from 'react';
+
+export const useImageUpload = () => {
+  const [selectedImage, setSelectedImage] = useState<File | null>(null);
+  const [imagePreview, setImagePreview] = useState<string | null>(null);
+
+  // 언마운트 시 이미지 URL 정리
+  useEffect(() => {
+    return () => {
+      if (imagePreview) {
+        URL.revokeObjectURL(imagePreview);
+      }
+    };
+  }, [imagePreview]);
+
+  // 이미지 선택 핸들러
+  const handleImageSelect = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const file = event.target.files?.[0];
+
+    if (file) {
+      setSelectedImage(file);
+      const previewUrl = URL.createObjectURL(file);
+      setImagePreview(previewUrl);
+    } else {
+      console.log('파일이 없음');
+    }
+  };
+
+  // 이미지 초기화
+  const clearImage = () => {
+    if (imagePreview) {
+      URL.revokeObjectURL(imagePreview);
+    }
+    setSelectedImage(null);
+    setImagePreview(null);
+  };
+
+  return {
+    selectedImage,
+    imagePreview,
+
+    handleImageSelect,
+    clearImage,
+    setSelectedImage,
+  };
+};
